+++ src/Endpoint/Update.php
@@ -28,4 +28,43 @@ class Update extends AbstractEndpoint
         
         return $this->client->request($payload);
     }
+    
+    /**
+     * @param string $body
+     * @param int $itemId
+     * @param int $parentId = null
+     * 
+     * @return array
+     */
+    public function create(string $body, int $itemId, int $parentId = null): array
+    {
+        $data = [
+            'body' => $body,
+            'item_id' => $itemId
+        ];
+        
+        if ($parentId !== null) {
+            $data['parent_id'] = $parentId;
+        }
+        
+        $payload = new Payload('mutation', [], [
+            new Payload('create_update ', $data, ['id'])
+        ]);
+        
+        return $this->client->request($payload);
+    }
+    
+    /**
+     * @param int $id
+     *
+     * @return array
+     */
+    public function delete(int $id): array
+    {
+        $payload = new Payload('mutation', [], [
+            new Payload('delete_update ', ['id' => $id], ['id'])
+        ]);
+        
+        return $this->client->request($payload);
+    }
 }